// src/components/interfaces.tsx
// Generated by https://quicktype.io
export interface Category {
    name: string;
    slug: string;
}

export interface Country {
    name: string;
    code: string;
}

export interface Tvg {
    id: string;
    name: string;
    url: string;
}

export interface Channel {
    name: string;
    logo: string;
    url: string;
    categories: Category[];
    countries: Country[];
    languages: Country[];
    tvg: Tvg;
}